// src/lib/api/generated/model/propertyResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Atuna API
 * A back-end for Atuna.
 * OpenAPI spec version: 1.0.0
 */
import type { PropertyResponseDescription } from "./propertyResponseDescription";
import type { PropertyResponsePropertyType } from "./propertyResponsePropertyType";
import type { PropertyResponseExpectedDownpayment } from "./propertyResponseExpectedDownpayment";
import type { PropertyResponseProjectName } from "./propertyResponseProjectName";
import type { PropertyResponseUpdatedAt } from "./propertyResponseUpdatedAt";

export interface PropertyResponse {
  title: string;
  description?: PropertyResponseDescription;
  price: string;
  location: string;
  property_type?: PropertyResponsePropertyType;
  expected_downpayment?: PropertyResponseExpectedDownpayment;
  project_name?: PropertyResponseProjectName;
  id: string;
  owner_id: string;
  status: string;
  is_archived: boolean;
  created_at: string;
  updated_at?: PropertyResponseUpdatedAt;
}